requirements.txt
Pillow
pyheif